refactor(#4912): extract footer-only marker helpers to keep mod.rs under the giant threshold
Move the self-contained footer-only marker rendering and accessors into a dedicated
submodule so task_notification_delivery/mod.rs remains below the 1000 production-line
giant threshold without changing marker behavior.
